Move update_iterator into shared context
