+++ src/systems/a320_systems/src/pneumatic.rs
@@ -12,7 +12,6 @@ use uom::si::{
 
 use systems::{
     accept_iterable,
-    hydraulic::update_iterator::MaxStepLoop,
     overhead::{AutoOffFaultPushButton, OnOffFaultPushButton},
     pneumatic::{
         valve::*, BleedMonitoringComputerChannelOperationMode,
@@ -23,9 +22,9 @@ use systems::{
         PressurizeableReservoir, TargetPressureSignal, VariableVolumeContainer,
     },
     shared::{
-        pid::PidController, ControllerSignal, ElectricalBusType, ElectricalBuses,
-        EngineCorrectedN1, EngineCorrectedN2, EngineFirePushButtons, HydraulicColor,
-        PneumaticValve, ReservoirAirPressure,
+        pid::PidController, update_iterator::MaxStepLoop, ControllerSignal, ElectricalBusType,
+        ElectricalBuses, EngineCorrectedN1, EngineCorrectedN2, EngineFirePushButtons,
+        HydraulicColor, PneumaticValve, ReservoirAirPressure,
     },
     simulation::{
         InitContext, Read, SimulationElement, SimulationElementVisitor, SimulatorReader,
